+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 ## [v0.0.2]
 
 - Add [#11]: Add flag for number of titles in a progress report line.
-- Add [#10]: Add tests using publid data from HathiTrust.
+- Add [#10]: Add tests using public data from HathiTrust.
 - Add [#7]: Save metainformation about titles.
 - Fix [#9]: Index is out of range in gnfinder.